suppression des dao dans les vues, separation de la mise à jour des eveenement (l'ajout/modif d'un support à part), revue de logique

+++ controllers/collaborator_crud.py
@@ -3,7 +3,7 @@ from DAO.collaborator_dao import CollaboratorDAO
 from DAO.department_dao import DepartmentDAO
 import bcrypt
 from datetime import datetime, timedelta, timezone
-from views.view_collaborator import view_create_collaborator, view_update_collaborator, wich_collaborator, view_delete_collaborator
+from views.view_collaborator import view_create_collaborator, view_update_collaborator, view_wich_collaborator, view_delete_collaborator
 
 
 session = create_db_connection()
@@ -12,7 +12,8 @@ department_dao = DepartmentDAO(session)
 
 def create_collaborator(token):
     created = False
-    info_collaborator = view_create_collaborator(created)
+    departments = department_dao.get_all_departments()
+    info_collaborator = view_create_collaborator(created, departments)
     if info_collaborator:
         password = info_collaborator[2]
         salt = bcrypt.gensalt()
@@ -29,27 +30,43 @@ def create_collaborator(token):
 
         if new_collaborator:
             created = True
-            view_create_collaborator(created)
+            view_create_collaborator(created, departments)
 
 
 def update_collaborator(token):
     collaborator = wich_collaborator()
+    departments = department_dao.get_all_departments()
+
     if collaborator:
         modified = False
-        new_data = view_update_collaborator(collaborator, modified)
+        new_data = view_update_collaborator(collaborator, modified, departments)
         print(new_data)
         if new_data:
             modification = collaborator_dao.update_collaborator(collaborator.id, new_data)
             if modification:
                 modified = True
-                view_update_collaborator(collaborator, modified)
+                view_update_collaborator(collaborator, modified, departments)
                 
 def delete_collaborator(token):
     collaborator = wich_collaborator()
     deleted = False
-    response = view_delete_collaborator(collaborator, deleted)
-    if response:
-        remove = collaborator_dao.delete_collaborator(collaborator.id)
-        if remove:
-            deleted = True
-            view_delete_collaborator(collaborator, deleted)
+    if collaborator:
+        response = view_delete_collaborator(collaborator, deleted)
+        if response:
+            remove = collaborator_dao.delete_collaborator(collaborator.id)
+            if remove:
+                deleted = True
+                view_delete_collaborator(collaborator, deleted)
+
+def wich_collaborator():
+    found = False
+    collaborators_corresponding = None
+    collaborator_name = view_wich_collaborator(collaborators_corresponding, found)
+    if collaborator_name:
+        collaborators_corresponding = collaborator_dao.get_corresponding_collaborator(collaborator_name)
+        if collaborators_corresponding:
+            found = True
+            collaborator = view_wich_collaborator(collaborators_corresponding, found)
+            if collaborator:
+                return collaborator
+
